backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6
websockets>=12.0
anthropic>=0.7.8
openai>=1.3.5
elevenlabs>=0.2.27
pdfplumber>=0.10.3
pydantic>=2.8.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
numpy>=1.26.0
pydub>=0.25.1
setuptools>=68.0.0
wheel
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

